Servicio de calcular prestamos, casi terminado

+++ models/calculo.go
@@ -9,6 +9,6 @@ definido como sql.Null, ejemplo sql.NullString. La mejor practica es evitar envi
 
 type Calculo struct {
 	Importe float64 `json:"importe"`
-	Meses   float64 `json:"meses"`
-	Interes float64 `json:"interes"`
+	Meses   int64   `json:"meses"`
+	Interes int64   `json:"interes"`
 }
